+++ stories/components/form-field/error.stories.ts
@@ -1,4 +1,4 @@
-import { Meta, StoryFn, StoryObj } from '@storybook/web-components';
+import { Meta, StoryObj } from '@storybook/web-components';
 import { html } from 'lit-html';
 
 const meta: Meta = {
@@ -8,5 +8,14 @@ const meta: Meta = {
 
 export default meta;
 
-const Template: StoryFn<Partial<any>> = () => html``;
-export const Default: StoryObj = Template.bind({});
+export const Default: StoryObj = {
+    render: () => html`
+        <wcs-form-field is-error="true">
+            <wcs-label>Full name</wcs-label>
+            <wcs-input placeholder="John Doe"></wcs-input>
+            <wcs-error>Your name is not valid, please do what is necessary
+                <a href="https://www.service-public.fr/particuliers/vosdroits/F1656">here</a>.
+            </wcs-error>
+        </wcs-form-field>
+    `
+}