+++ stories/components/form-field/label.stories.ts
@@ -1,4 +1,4 @@
-import { Meta, StoryFn, StoryObj } from '@storybook/web-components';
+import { Meta, StoryObj } from '@storybook/web-components';
 import { html } from 'lit-html';
 
 const meta: Meta = {
@@ -8,5 +8,14 @@ const meta: Meta = {
 
 export default meta;
 
-const Template: StoryFn<Partial<any>> = () => html``;
-export const Default: StoryObj = Template.bind({});
+export const Default: StoryObj = {
+    render: (args) => html`
+        <wcs-form-field>
+            <wcs-label .required=${args.required}>Full name</wcs-label>
+            <wcs-input .required=${args.required} placeholder="John Doe"></wcs-input>
+        </wcs-form-field>
+    `,
+    args: {
+        required: false,
+    }
+}